Enabled the Passing of Additional CLI Arguments
It is now possible to pass additional command line arguments to ComfyUI
in the "docker run" command. This was done by moving the "CMD" from the
Dockerfile to the "entrypoint.sh" script. Now, users can specify
additional command line arguments when starting the container as the
command, which are then passed to ComfyUI. The read me was updated to
showcase this.

Also, the changelog was updated to reflect this change, as well as the
changes introduced by the previous PR. The list of contributors was
updated to include the contributor that suggested this change.

Finally, in the read me, the commands for building and running the image
locally had a small issue: After cloning the repository, the current
directory was not changed to the cloned repository, which would lead to
an error when building the image. This was fixed by adding a
"cd comfyui-docker" command after the "git clone" command.